+++ frontend/src/app/components/modal/manage-doc-types/AddDocTypeModal.tsx
@@ -1,51 +1,37 @@
-import { FC, useEffect, useState } from 'react';
+import { FC, useState } from 'react';
 import { Button, Col, Form, Modal, Row, Spinner } from 'react-bootstrap';
-import { DocType } from '../../../../types/types';
+import { DocType } from '../../../types/types';
 
-interface EditDocTypeModalProps {
-  documentType: DocType;
+interface AddDocTypeModalProps {
   allDocTypes: DocType[];
   show: boolean;
   onHide: () => void;
-  onEdit: (id: number, name: string, created_by: string, created_date: string) => void;
+  onAdd: (name: string, created_by: string, created_date: string) => void;
 }
 
-const EditDocTypeModal: FC<EditDocTypeModalProps> = ({ documentType, allDocTypes, show, onHide, onEdit }) => {
+const AddDocTypeModal: FC<AddDocTypeModalProps> = ({ allDocTypes, show, onHide, onAdd }) => {
   const [name, setName] = useState<string>('');
   const [createdBy, setCreatedBy] = useState<string>('');
-  const [createdDate, setCreatedDate] = useState<string>('');
-  const [defaultDisabled, setDefaultDisabled] = useState<boolean>(true);
+  const [createdDate, setCreatedDate] = useState<string>(new Date().toISOString().substring(0, 10));
   const [loading, setLoading] = useState<boolean>(false);
   const [error, setError] = useState<string>('');
   const [showError, setShowError] = useState<boolean>(false);
+  const [defaultDisabled, setDefaultDisabled] = useState<boolean>(true);
 
-  useEffect(() => {
-    setName(documentType.name);
-    setCreatedBy(documentType.created_by);
-    setCreatedDate(documentType.created_date?.substring(0, 10));
-  }, [documentType]);
-
-  const editDocTypeHandler = async () => {
+  const addDocTypeHandler = async () => {
     try {
       setLoading(true);
       setShowError(false);
-      const sameDocTypes = allDocTypes.filter((docType) => {
-        return docType.name === name;
-      });
-      const isNameUnique = sameDocTypes.length <= 1; // there should only be up to one docType (the current one) that has the same name
-      const isNameNotEmpty = name.length !== 0;
-      console.log('name: ' + name);
-      console.log(sameDocTypes);
-      console.log(isNameUnique);
-      if (isNameUnique && isNameNotEmpty) {
-        onEdit(documentType.id, name, createdBy, createdDate);
+      const isNameUnique = !allDocTypes.some((docType) => docType.name === name);
+      if (isNameUnique) {
+        onAdd(name, createdBy, createdDate);
         onHide();
       } else {
         setError('That document type name already exists');
         setShowError(true);
       }
     } catch (err) {
-      setError('Error updating Document Type');
+      setError('Error adding new Document Type');
       setShowError(true);
       console.log(err);
     } finally {
@@ -59,28 +45,28 @@ const EditDocTypeModal: FC<EditDocTypeModalProps> = ({ documentType, allDocTypes
   };
 
   const handleCreatedByChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (defaultDisabled) setDefaultDisabled(false);
+    console.log(e.target.value);
     setCreatedBy(e.target.value);
   };
 
   const handleCreatedDateChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (defaultDisabled) setDefaultDisabled(false);
+    console.log(e.target.value);
     setCreatedDate(e.target.value);
   };
 
   return (
     <Modal show={show} onHide={onHide} size="lg">
       <Modal.Header closeButton>
-        <Modal.Title>Edit Document Type</Modal.Title>
+        <Modal.Title>Add Document Type</Modal.Title>
       </Modal.Header>
       <Modal.Body>
         <Form>
           <Form.Group as={Row} className="mb-3">
             <Form.Label column sm="10">
-              Name:
+              Document Type Name:
             </Form.Label>
             <Col sm="10">
-              <Form.Control type="text" defaultValue={documentType.name} onChange={handleNameChange} />
+              <Form.Control type="text" onChange={handleNameChange} />
             </Col>
           </Form.Group>
           <Form.Group as={Row} className="mb-3">
@@ -88,7 +74,7 @@ const EditDocTypeModal: FC<EditDocTypeModalProps> = ({ documentType, allDocTypes
               Created By:
             </Form.Label>
             <Col sm="10">
-              <Form.Control type="text" defaultValue={documentType.created_by} onChange={handleCreatedByChange} />
+              <Form.Control type="text" onChange={handleCreatedByChange} />
             </Col>
           </Form.Group>
           <Form.Group as={Row} className="mb-3">
@@ -107,12 +93,12 @@ const EditDocTypeModal: FC<EditDocTypeModalProps> = ({ documentType, allDocTypes
           Cancel
         </Button>
 
-        <Button variant="primary" onClick={editDocTypeHandler} disabled={loading || defaultDisabled}>
-          {loading ? <Spinner as="span" animation="border" size="sm" role="status" aria-hidden="true" /> : 'Update'}
+        <Button variant="primary" onClick={addDocTypeHandler} disabled={loading || !name || defaultDisabled}>
+          {loading ? <Spinner as="span" animation="border" size="sm" role="status" aria-hidden="true" /> : 'Add'}
         </Button>
       </Modal.Footer>
     </Modal>
   );
 };
 
-export default EditDocTypeModal;
+export default AddDocTypeModal;
