+++ frontend/src/app/components/modal/manage-provisions/RemoveProvisionModal.tsx
@@ -1,27 +1,29 @@
 import { FC, useState } from 'react';
 import { Button, Form, Modal, Row, Spinner } from 'react-bootstrap';
-import { DocType } from '../../../../types/types';
+import { Provision } from '../../../types/types';
 
-interface RemoveDocTypeModalProps {
-  documentType: DocType;
+interface RemoveProvisionModalProps {
+  provision: Provision | undefined;
   show: boolean;
   onHide: () => void;
   onRemove: (id: number) => void;
 }
 
-const RemoveDocTypeModal: FC<RemoveDocTypeModalProps> = ({ documentType, show, onHide, onRemove }) => {
+const RemoveProvisionModal: FC<RemoveProvisionModalProps> = ({ provision, show, onHide, onRemove }) => {
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState('');
   const [showError, setShowError] = useState(false);
 
-  const removeDocTypeHandler = async () => {
+  const removeProvisionHandler = async () => {
     try {
-      setLoading(true);
-      setShowError(false);
-      onRemove(documentType.id);
-      onHide();
+      if (provision) {
+        setLoading(true);
+        setShowError(false);
+        onRemove(provision.id);
+        onHide();
+      }
     } catch (err) {
-      setError('Error removing Document Type');
+      setError('Error removing provision');
       setShowError(true);
       console.log(err);
     } finally {
@@ -36,10 +38,10 @@ const RemoveDocTypeModal: FC<RemoveDocTypeModalProps> = ({ documentType, show, o
       </Modal.Header>
       <Modal.Body>
         <Form>
-          <p>Are you sure you want to remove this document type?</p>
+          <p>Are you sure you want to remove this provision?</p>
           <Form.Group as={Row} className="mb-3">
             <Form.Label column sm="10">
-              Document Type: <b>{documentType.name}</b>
+              <b>{provision?.provision_name}</b>
             </Form.Label>
           </Form.Group>
         </Form>
@@ -50,7 +52,7 @@ const RemoveDocTypeModal: FC<RemoveDocTypeModalProps> = ({ documentType, show, o
           Cancel
         </Button>
 
-        <Button variant="primary" onClick={removeDocTypeHandler} disabled={loading}>
+        <Button variant="primary" onClick={removeProvisionHandler} disabled={loading}>
           {loading ? <Spinner as="span" animation="border" size="sm" role="status" aria-hidden="true" /> : 'Remove'}
         </Button>
       </Modal.Footer>
@@ -58,4 +60,4 @@ const RemoveDocTypeModal: FC<RemoveDocTypeModalProps> = ({ documentType, show, o
   );
 };
 
-export default RemoveDocTypeModal;
+export default RemoveProvisionModal;
