+++ frontend/src/app/components/modal/manage-provisions/forms/EditProvisionModalForm.tsx
@@ -1,27 +1,35 @@
 import { useEffect, useState } from 'react';
-import { DocType, GroupMax, ProvisionUpload } from '../../../../types/types';
+import { DocType, GroupMax, Provision, ProvisionUpload, Variable } from '../../../../types/types';
 import { Button, Col, Form, Modal, Row, Spinner } from 'react-bootstrap';
+import ManageVariablesTable from '../../../table/manage-provisions/ManageVariablesTable';
 
-interface AddProvisionModalProps {
-  groupMaxArray: GroupMax[] | undefined;
-  show: boolean;
+interface EditProvisionModalFormProps {
+  provision: Provision | undefined;
+  variables: Variable[] | undefined;
   documentTypes: DocType[] | undefined;
-  addProvisionHandler: (provision: ProvisionUpload) => void;
+  groupMaxArray: GroupMax[] | undefined;
+  loading: boolean;
+  updateProvisionHandler: (provision: ProvisionUpload, provisionId: number) => void;
   onHide: () => void;
-  refreshTables: () => void;
+  onDisplayAdd: () => void;
+  onDisplayEdit: (variableId: number) => void;
+  onDisplayRemove: (variableId: number) => void;
 }
 
-const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
-  groupMaxArray,
-  show,
+const EditProvisionModalForm: React.FC<EditProvisionModalFormProps> = ({
+  provision,
+  variables,
   documentTypes,
-  addProvisionHandler,
+  loading,
+  groupMaxArray,
+  updateProvisionHandler,
   onHide,
-  refreshTables,
+  onDisplayAdd,
+  onDisplayEdit,
+  onDisplayRemove,
 }) => {
-  const [loading, setLoading] = useState<boolean>(false);
   const [type, setType] = useState<string>('M');
-  const [group, setGroup] = useState<number>(1);
+  const [group, setGroup] = useState<number | null>(null);
   const [groupDescription, setGroupDescription] = useState<string>('');
   const [max, setMax] = useState<number | null>(null);
   const [maxInputValue, setMaxInputValue] = useState<string>('');
@@ -30,8 +38,30 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
   const [freeText, setFreeText] = useState<string>('');
   const [helpText, setHelpText] = useState<string>('');
   const [category, setCategory] = useState<string>('');
-  const [order, setOrder] = useState<number>(1);
+  const [order, setOrder] = useState<number | undefined>();
   const [documentTypeIds, setDocumentTypeIds] = useState<number[]>([]);
+
+  useEffect(() => {
+    const getData = async () => {
+      if (provision && groupMaxArray) {
+        setType(provision.type);
+        setGroup(provision.provision_group);
+        if (groupMaxArray) {
+          const groupMax = groupMaxArray.find((group) => group.provision_group === provision.provision_group);
+          setGroupDescription(groupMax ? groupMax.provision_group_text : '');
+        }
+        setMax(provision.max);
+        setProvisionName(provision.provision_name);
+        setFreeText(provision.free_text);
+        setHelpText(provision.help_text);
+        setCategory(provision.category);
+        setOrder(provision.order_value);
+        provision.document_type_ids ? setDocumentTypeIds(provision.document_type_ids) : setDocumentTypeIds([]);
+      }
+    };
+
+    getData();
+  }, [provision, groupMaxArray]);
 
   useEffect(() => {
     setIsMaxUnlimited(max ? max >= 999 : false);
@@ -114,64 +144,46 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
     if (e.target.checked) {
       setDocumentTypeIds((prevDocTypeIds) => [...prevDocTypeIds, documentTypeId]);
     } else {
-      setDocumentTypeIds(documentTypeIds.filter((v) => v === documentTypeId));
+      setDocumentTypeIds((prevDocTypeIds) => prevDocTypeIds.filter((id) => id !== documentTypeId));
     }
   };
 
   const handleSaveButton = () => {
-    try {
-      setLoading(true);
-      if (max) {
-        const trueMax = isMaxUnlimited ? 999 : max;
-        const provisionUpload: ProvisionUpload = {
-          type: type,
-          provision_group: group,
-          provision_group_text: groupDescription,
-          max: trueMax,
-          provision_name: provisionName,
-          free_text: freeText,
-          help_text: helpText,
-          category: category,
-          order_value: order,
-          document_type_ids: documentTypeIds,
-        };
-        addProvisionHandler(provisionUpload);
-        onHide();
-        refreshTables();
-      }
-    } catch (err) {
-      console.log('Error adding provision');
-      console.log(err);
-    } finally {
-      setLoading(false);
+    if (provision && max) {
+      const trueMax = isMaxUnlimited ? 999 : max;
+      const provisionUpload: ProvisionUpload = {
+        type: type,
+        provision_group: group ? group : 1,
+        provision_group_text: groupDescription,
+        max: trueMax,
+        provision_name: provisionName,
+        free_text: freeText,
+        help_text: helpText,
+        category: category,
+        order_value: order ? order : 1,
+        document_type_ids: documentTypeIds,
+      };
+      updateProvisionHandler(provisionUpload, provision.id);
     }
   };
 
+  const isDocumentTypeChecked = (documentTypeId: number): boolean => {
+    return documentTypeIds.find((d) => d === documentTypeId) ? true : false;
+  };
+
   return (
-    <Modal show={show} onHide={onHide} size="lg">
-      <Modal.Header>
-        <Modal.Title>Add Provision</Modal.Title>
-        <Button
-          variant="none"
-          onClick={onHide}
-          style={{
-            marginLeft: 'auto',
-            border: 'none',
-            backgroundColor: 'transparent',
-            color: 'black',
-          }}
-        >
-          &times;
-        </Button>
+    <>
+      <Modal.Header closeButton>
+        <Modal.Title>Edit Provision</Modal.Title>
       </Modal.Header>
       <Modal.Body>
         <Form>
-          <Form.Group controlId="searchEmail" className="mb-3">
+          <Form.Group className="mb-3">
             <Form.Label column sm={2}>
               Type:
             </Form.Label>
             <Col sm={12}>
-              <Form.Select value="O" className="form-control" onChange={handleTypeChange}>
+              <Form.Select value={type} className="form-control" onChange={handleTypeChange}>
                 <option value="O">O</option>
                 <option value="M">M</option>
                 <option value="B">B</option>
@@ -183,7 +195,12 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
               Group
             </Form.Label>
             <Col sm="12">
-              <Form.Control type="number" name="provision_group" onChange={handleGroupChange} />
+              <Form.Control
+                type="number"
+                defaultValue={group ? group : ''}
+                name="provision_group"
+                onChange={handleGroupChange}
+              />
             </Col>
           </Form.Group>
 
@@ -194,8 +211,8 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
             <Col sm="12">
               <Form.Control
                 type="text"
+                defaultValue={groupDescription}
                 name="provision_group_text"
-                value={groupDescription}
                 onChange={handleGroupDescriptionChange}
               />
             </Col>
@@ -208,8 +225,8 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
             <Col sm="12">
               <Form.Control
                 type="number"
-                name="max"
                 value={maxInputValue}
+                name="max"
                 onChange={handleMaxChange}
                 disabled={isMaxUnlimited}
               />
@@ -230,7 +247,12 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
               Provision
             </Form.Label>
             <Col sm="10">
-              <Form.Control type="text" name="provision" onChange={handleProvisionNameChange} />
+              <Form.Control
+                type="text"
+                defaultValue={provisionName}
+                name="provision"
+                onChange={handleProvisionNameChange}
+              />
             </Col>
           </Form.Group>
 
@@ -239,7 +261,13 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
               Free Text
             </Form.Label>
             <Col sm="10">
-              <Form.Control as="textarea" rows={3} name="free_text" onChange={handleFreeTextChange} />
+              <Form.Control
+                as="textarea"
+                rows={3}
+                name="free_text"
+                defaultValue={freeText}
+                onChange={handleFreeTextChange}
+              />
             </Col>
           </Form.Group>
 
@@ -248,7 +276,7 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
               Help Text
             </Form.Label>
             <Col sm="12">
-              <Form.Control type="text" name="help_text" onChange={handleHelpTextChange} />
+              <Form.Control type="text" defaultValue={helpText} name="help_text" onChange={handleHelpTextChange} />
             </Col>
           </Form.Group>
 
@@ -257,9 +285,10 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
               Category
             </Form.Label>
             <Col sm="12">
-              <Form.Control type="text" name="category" onChange={handleCategoryTextChange} />
+              <Form.Control type="text" defaultValue={category} name="category" onChange={handleCategoryTextChange} />
             </Col>
           </Form.Group>
+
           <Form.Group className="mb-3">
             <Form.Label column sm="12">
               Order
@@ -282,6 +311,7 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
                 <Col sm={{ span: 4 }}>
                   <Form.Check
                     name={`document-type-${docType.id}`}
+                    checked={isDocumentTypeChecked(docType.id)}
                     value={docType.id}
                     onChange={handleDocumentTypeIdUpdate}
                   />
@@ -289,17 +319,32 @@ const AddProvisionModal: React.FC<AddProvisionModalProps> = ({
               </Form.Group>
             );
           })}
+        <Col sm="12">
+          <Form.Label style={{ marginTop: '10px' }}>Variables</Form.Label>
+        </Col>
+        <Col sm="12">
+          <ManageVariablesTable
+            variables={variables ? variables : []}
+            onDisplayEdit={onDisplayEdit}
+            onDisplayRemove={onDisplayRemove}
+          />
+        </Col>
+        <Col sm="12">
+          <Button onClick={onDisplayAdd} variant="success">
+            Add a Variable
+          </Button>
+        </Col>
       </Modal.Body>
       <Modal.Footer>
         <Button variant="secondary" onClick={onHide}>
           Cancel
         </Button>
 
-        <Button variant="primary" onClick={handleSaveButton} disabled={loading}>
+        <Button variant="primary" onClick={handleSaveButton} disabled={loading || !provision}>
           {loading ? <Spinner as="span" animation="border" size="sm" role="status" aria-hidden="true" /> : 'Save'}
         </Button>
       </Modal.Footer>
-    </Modal>
+    </>
   );
 };
 
@@ -326,4 +371,4 @@ const FormLabelWithPeriods: React.FC<{ text: string }> = ({ text }) => {
   );
 };
 
-export default AddProvisionModal;
+export default EditProvisionModalForm;
