+++ frontend/src/app/components/table/manage-doc-types/EditDocTypeTable.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+interface EditDocTypeTableProps {}
+
+const EditDocTypeTable: FC<EditDocTypeTableProps> = () => {
+  return <></>;
+};
+
+export default EditDocTypeTable;
