+++ frontend/src/app/components/table/manage-doc-types/ManageDocTypesTable.tsx
