+++ frontend/src/app/components/table/manage-doc-types/ProvisionGroupsTable.tsx
@@ -1,0 +1,101 @@
+import { FC, useEffect, useState } from 'react';
+import { ProvisionGroup } from '../../../types/types';
+import { Column, ColumnDef, ColumnMeta, Row, Table, createColumnHelper } from '@tanstack/react-table';
+import LinkButton from '../../common/LinkButton';
+import { DataTable } from '../common/DataTable';
+
+interface ProvisionGroupsTableProps {
+  provisionGroups: ProvisionGroup[];
+  onUpdate: (provisionGroups: ProvisionGroup[]) => void;
+  showRemove: (provisionGroup: ProvisionGroup) => void;
+}
+
+const ProvisionGroupsTable: FC<ProvisionGroupsTableProps> = ({ provisionGroups, onUpdate, showRemove }) => {
+  const [updatedProvisionGroups, setUpdatedProvisionGroups] = useState<ProvisionGroup[]>(provisionGroups);
+
+  useEffect(() => {
+    setUpdatedProvisionGroups(provisionGroups);
+  }, [provisionGroups]);
+
+  const showRemoveHandler = (provisionGroup: ProvisionGroup) => {
+    showRemove(provisionGroup);
+  };
+
+  const handleCellUpdate = (rowIndex: number, columnId: keyof ProvisionGroup, newValue: any) => {
+    const newProvisionGroups = updatedProvisionGroups.map((group, index) => {
+      if (index === rowIndex) {
+        const updatedValue = columnId === 'provision_group' || columnId === 'max' ? parseInt(newValue, 10) : newValue;
+
+        return { ...group, [columnId]: updatedValue };
+      }
+      return group;
+    });
+
+    setUpdatedProvisionGroups(newProvisionGroups);
+    onUpdate(newProvisionGroups);
+  };
+
+  const columnHelper = createColumnHelper<ProvisionGroup>();
+
+  const columns: ColumnDef<ProvisionGroup, any>[] = [
+    columnHelper.accessor('provision_group', {
+      id: 'provision_group',
+      cell: (info) => (
+        <TableCell getValue={info.getValue} row={info.row} column={info.column} onCellUpdate={handleCellUpdate} />
+      ),
+
+      header: () => 'Group',
+      meta: { customCss: { width: '12%' }, type: 'text' },
+    }),
+    columnHelper.accessor('provision_group_text', {
+      id: 'provision_group_text',
+      cell: (info) => (
+        <TableCell getValue={info.getValue} row={info.row} column={info.column} onCellUpdate={handleCellUpdate} />
+      ),
+      header: () => 'Description',
+      meta: { customCss: { width: '64%', margin: '0px' }, type: 'text' },
+    }),
+    columnHelper.accessor('max', {
+      id: 'max',
+      cell: (info) => (
+        <TableCell getValue={info.getValue} row={info.row} column={info.column} onCellUpdate={handleCellUpdate} />
+      ),
+      header: () => 'Max',
+      meta: { customCss: { width: '12%' }, type: 'text' },
+    }),
+    columnHelper.display({
+      id: 'remove',
+      cell: (info) => <LinkButton text="Remove" onClick={() => showRemoveHandler(info.row.original)} />,
+      header: () => null,
+      meta: { customCss: { width: '12%' } },
+    }),
+  ];
+
+  return <DataTable columns={columns} data={updatedProvisionGroups} />;
+};
+export default ProvisionGroupsTable;
+
+interface TableCellProps<T> {
+  getValue: () => any;
+  row: Row<T>;
+  column: Column<T, any>;
+  onCellUpdate: (rowIndex: number, columnId: keyof ProvisionGroup, newValue: any) => void; // Updated type
+}
+
+const TableCell: FC<TableCellProps<ProvisionGroup>> = ({ getValue, row, column, onCellUpdate }) => {
+  const initialValue = getValue();
+  const [value, setValue] = useState(initialValue);
+
+  useEffect(() => {
+    setValue(initialValue);
+  }, [initialValue]);
+
+  const onBlur = () => {
+    const validKeys: (keyof ProvisionGroup)[] = ['id', 'provision_group', 'provision_group_text', 'max'];
+    if (validKeys.includes(column.id as keyof ProvisionGroup)) {
+      onCellUpdate(row.index, column.id as keyof ProvisionGroup, value);
+    }
+  };
+
+  return <input value={value} onChange={(e) => setValue(e.target.value)} onBlur={onBlur} type={'text'} />;
+};
