+++ frontend/src/app/components/table/manage-provisions/ManageProvisionsTable.tsx
