+++ frontend/src/app/components/table/manage-provisions/ManageVariablesTable.tsx
